Users can ignore friend requests
requests now have an  ignored field. When rejected, a requests ignored value is set to TRUE (1) and the request is no longer shown in the users friend request screen.

Note: This is better than simply deleting the requests as then the request could be sent repeatedly. This system prevents multiple requests.